Zarr: implement deferred driver loading
