+++ frmts/zarr/CMakeLists.txt
@@ -11,9 +11,16 @@ add_gdal_driver(
           zarr_v3_group.cpp
           zarr_sharedresource.cpp
           zarrdriver.cpp
-          STRONG_CXX_WFLAGS
-          PLUGIN_CAPABLE
-          NO_DEPS)
+  CORE_SOURCES
+          zarrdrivercore.cpp
+  STRONG_CXX_WFLAGS
+  PLUGIN_CAPABLE
+  NO_DEPS)
+
+if(NOT TARGET gdal_Zarr)
+    return()
+endif()
+
 gdal_standard_includes(gdal_Zarr)
 target_include_directories(gdal_Zarr PRIVATE ${GDAL_RASTER_FORMAT_SOURCE_DIR}/mem
                                              ${GDAL_RASTER_FORMAT_SOURCE_DIR}/netcdf)
