+++ frmts/zarr/zarrdriver.cpp
@@ -27,6 +27,7 @@
  ****************************************************************************/
 
 #include "zarr.h"
+#include "zarrdrivercore.h"
 
 #include "cpl_minixml.h"
 
@@ -45,51 +46,6 @@
 ZarrDataset::ZarrDataset(const std::shared_ptr<GDALGroup> &poRootGroup)
     : m_poRootGroup(poRootGroup)
 {
-}
-
-/************************************************************************/
-/*                    CheckExistenceOfOneZarrFile()                     */
-/************************************************************************/
-
-static bool CheckExistenceOfOneZarrFile(const char *pszFilename)
-{
-
-    CPLString osMDFilename = CPLFormFilename(pszFilename, ".zarray", nullptr);
-
-    VSIStatBufL sStat;
-    if (VSIStatL(osMDFilename, &sStat) == 0)
-        return true;
-
-    osMDFilename = CPLFormFilename(pszFilename, ".zgroup", nullptr);
-    if (VSIStatL(osMDFilename, &sStat) == 0)
-        return true;
-
-    // Zarr V3
-    osMDFilename = CPLFormFilename(pszFilename, "zarr.json", nullptr);
-    if (VSIStatL(osMDFilename, &sStat) == 0)
-        return true;
-
-    return false;
-}
-
-/************************************************************************/
-/*                              Identify()                              */
-/************************************************************************/
-
-int ZarrDataset::Identify(GDALOpenInfo *poOpenInfo)
-
-{
-    if (STARTS_WITH(poOpenInfo->pszFilename, "ZARR:"))
-    {
-        return TRUE;
-    }
-
-    if (!poOpenInfo->bIsDirectory)
-    {
-        return FALSE;
-    }
-
-    return CheckExistenceOfOneZarrFile(poOpenInfo->pszFilename);
 }
 
 /************************************************************************/
@@ -284,7 +240,7 @@ GetExtraDimSampleCount(const std::shared_ptr<GDALMDArray> &poArray,
 
 GDALDataset *ZarrDataset::Open(GDALOpenInfo *poOpenInfo)
 {
-    if (!Identify(poOpenInfo))
+    if (!ZARRDriverIdentify(poOpenInfo))
     {
         return nullptr;
     }
@@ -1715,56 +1671,12 @@ CPLErr ZarrRasterBand::IRasterIO(GDALRWFlag eRWFlag, int nXOff, int nYOff,
 void GDALRegister_Zarr()
 
 {
-    if (GDALGetDriverByName("Zarr") != nullptr)
+    if (GDALGetDriverByName(DRIVER_NAME) != nullptr)
         return;
 
     GDALDriver *poDriver = new ZarrDriver();
+    ZARRDriverSetCommonMetadata(poDriver);
 
-    poDriver->SetDescription("Zarr");
-    poDriver->SetMetadataItem(GDAL_DCAP_RASTER, "YES");
-    poDriver->SetMetadataItem(GDAL_DCAP_MULTIDIM_RASTER, "YES");
-    poDriver->SetMetadataItem(GDAL_DMD_LONGNAME, "Zarr");
-    poDriver->SetMetadataItem(GDAL_DMD_CREATIONDATATYPES,
-                              "Byte Int16 UInt16 Int32 UInt32 Int64 UInt64 "
-                              "Float32 Float64 CFloat32 CFloat64");
-    poDriver->SetMetadataItem(GDAL_DCAP_VIRTUALIO, "YES");
-    poDriver->SetMetadataItem(GDAL_DMD_SUBDATASETS, "YES");
-
-    poDriver->SetMetadataItem(
-        GDAL_DMD_OPENOPTIONLIST,
-        "<OpenOptionList>"
-        "   <Option name='USE_ZMETADATA' type='boolean' description='Whether "
-        "to use consolidated metadata from .zmetadata' default='YES'/>"
-        "   <Option name='CACHE_TILE_PRESENCE' type='boolean' "
-        "description='Whether to establish an initial listing of present "
-        "tiles' default='NO'/>"
-        "   <Option name='MULTIBAND' type='boolean' default='YES' "
-        "description='Whether to expose >= 3D arrays as GDAL multiband "
-        "datasets "
-        "(when using the classic 2D API)'/>"
-        "   <Option name='DIM_X' type='string' description="
-        "'Name or index of the X dimension (only used when MULTIBAND=YES)'/>"
-        "   <Option name='DIM_Y' type='string' description="
-        "'Name or index of the Y dimension (only used when MULTIBAND=YES)'/>"
-        "   <Option name='LOAD_EXTRA_DIM_METADATA_DELAY' type='string' "
-        "description="
-        "'Maximum delay in seconds allowed to set the DIM_{dimname}_VALUE band "
-        "metadata items'/>"
-        "</OpenOptionList>");
-
-    poDriver->SetMetadataItem(
-        GDAL_DMD_MULTIDIM_DATASET_CREATIONOPTIONLIST,
-        "<MultiDimDatasetCreationOptionList>"
-        "   <Option name='FORMAT' type='string-select' default='ZARR_V2'>"
-        "     <Value>ZARR_V2</Value>"
-        "     <Value>ZARR_V3</Value>"
-        "   </Option>"
-        "   <Option name='CREATE_ZMETADATA' type='boolean' "
-        "description='Whether to create consolidated metadata into .zmetadata "
-        "(Zarr V2 only)' default='YES'/>"
-        "</MultiDimDatasetCreationOptionList>");
-
-    poDriver->pfnIdentify = ZarrDataset::Identify;
     poDriver->pfnOpen = ZarrDataset::Open;
     poDriver->pfnCreateMultiDimensional = ZarrDataset::CreateMultiDimensional;
     poDriver->pfnCreate = ZarrDataset::Create;
