+++ gcore/gdal_frmts.h
@@ -222,6 +222,7 @@ void CPL_DLL GDALRegister_TGA(void);
 void CPL_DLL GDALRegister_OGCAPI(void);
 void CPL_DLL GDALRegister_STACTA(void);
 void CPL_DLL GDALRegister_Zarr(void);
+void DeclareDeferredZarrPlugin(void);
 void CPL_DLL GDALRegister_STACIT(void);
 void CPL_DLL GDALRegister_JPEGXL(void);
 void DeclareDeferredJPEGXLPlugin(void);
